Shuffle checkValidationCriteria
